+++ src/NEventStore.PollingClientExample/MainProgram.cs
@@ -2,7 +2,7 @@ namespace NEventStore.PollingClientExample
 {
     using System;
     using NEventStore.PollingClient;
-    using NEventStore.Logging;
+    using Microsoft.Extensions.Logging;
 
     internal static class MainProgram
     {
@@ -51,9 +51,16 @@ namespace NEventStore.PollingClientExample
 
         private static IStoreEvents WireupEventStore()
         {
+            var loggerFactory = LoggerFactory.Create(logging =>
+            {
+                logging
+                    .AddConsole()
+                    .AddDebug()
+                    .SetMinimumLevel(Microsoft.Extensions.Logging.LogLevel.Trace);
+            });
+
             return Wireup.Init()
-               .LogToOutputWindow(LogLevel.Verbose)
-               .LogToConsoleWindow(LogLevel.Verbose)
+               .WithLoggerFactory(loggerFactory)
                .UsingInMemoryPersistence()
                .InitializeStorageEngine()
 #if NET461